requirements.txt
Django
mysqlclient
openpyxl
tablib
django-import-export
geopy
django-crispy-forms
crispy-bootstrap4
Pillow
matplotlib
seaborn
django-bootstrap4
gunicorn
djangorestframework
reportlab
